Move geocoding helpers to GeoCoder class
