+++ geode-core/src/main/java/org/apache/geode/redis/internal/Coder.java
@@ -14,12 +14,6 @@
  */
 package org.apache.geode.redis.internal;
 
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufAllocator;
-import org.apache.geode.cache.EntryDestroyedException;
-import org.apache.geode.cache.query.Struct;
-import org.apache.geode.redis.internal.org.apache.hadoop.fs.GeoCoord;
-
 import java.io.UnsupportedEncodingException;
 import java.text.DecimalFormat;
 import java.util.Arrays;
@@ -28,6 +22,13 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufAllocator;
+
+import org.apache.geode.cache.EntryDestroyedException;
+import org.apache.geode.cache.query.Struct;
+import org.apache.geode.redis.internal.org.apache.hadoop.fs.GeoCoord;
 
 /**
  * This is a safe encoder and decoder for all redis matching needs
@@ -94,7 +95,6 @@ public class Coder {
   public static final String CHARSET = "UTF-8";
 
   protected static final DecimalFormat decimalFormatter = new DecimalFormat("#");
-  public static final int LEN_GEOHASH = 60;
 
   static {
     decimalFormatter.setMaximumFractionDigits(10);
@@ -330,7 +330,8 @@ public class Coder {
     return buf;
   }
 
-  public static ByteBuf getBulkStringGeoCoordinateArrayResponse(ByteBufAllocator alloc, Collection<GeoCoord> items) {
+  public static ByteBuf getBulkStringGeoCoordinateArrayResponse(ByteBufAllocator alloc,
+      Collection<GeoCoord> items) {
     Iterator<GeoCoord> it = items.iterator();
     ByteBuf response = alloc.buffer();
     response.writeByte(Coder.ARRAY_ID);
@@ -342,11 +343,9 @@ public class Coder {
         tmp.writeBytes(Coder.bNIL);
       } else {
         tmp.writeBytes(Coder.getBulkStringArrayResponse(alloc,
-                Arrays.asList(
-                        Double.toString(next.getLongitude()),
-                        Double.toString(next.getLatitude())
-                )
-        ));
+            Arrays.asList(
+                Double.toString(next.getLongitude()),
+                Double.toString(next.getLatitude()))));
       }
       size++;
     }
@@ -360,7 +359,8 @@ public class Coder {
     return response;
   }
 
-  public static ByteBuf getBulkStringArrayResponseOfValues(ByteBufAllocator alloc, Collection<?> items) {
+  public static ByteBuf getBulkStringArrayResponseOfValues(ByteBufAllocator alloc,
+      Collection<?> items) {
     Iterator<?> it = items.iterator();
     ByteBuf response = alloc.buffer();
     response.writeByte(Coder.ARRAY_ID);
@@ -547,133 +547,6 @@ public class Coder {
       return Double.NEGATIVE_INFINITY;
     else
       return Double.parseDouble(d);
-  }
-
-  public static GeoCoord geoPos(String hash) {
-    StringBuilder binStringBuilder = new StringBuilder();
-    for (char digit : hash.toCharArray()) {
-      int val = base32Val(digit);
-      binStringBuilder.append(base32bin(val));
-    }
-
-    char[] binChars = binStringBuilder.toString().toCharArray();
-    char[] lonChars = new char[binChars.length/2];
-    char[] latChars = new char[binChars.length/2];
-    for (int i = 0; i < binChars.length; i+=2) {
-      lonChars[i/2] = binChars[i];
-      latChars[i/2] = binChars[i+1];
-    }
-
-    return new GeoCoord(coord(lonChars, -180.0, 180.0), coord(latChars, -90.0, 90.0));
-  }
-
-  public static Double geoDist(String hash1, String hash2) {
-    GeoCoord coord1 = geoPos(hash1);
-    GeoCoord coord2 = geoPos(hash2);
-
-    Double lat1 = Math.toRadians(coord1.getLatitude());
-    Double long1 = Math.toRadians(coord1.getLongitude());
-    Double lat2 = Math.toRadians(coord2.getLatitude());
-    Double long2 = Math.toRadians(coord2.getLongitude());
-
-    Double hav = haversine(lat2-lat1) + (Math.cos(lat1) * Math.cos(lat2) * haversine(long2-long1));
-    Double distAngle = Math.acos(1 - (2 * hav));
-
-    return 6372797.560856 * distAngle;
-  }
-
-  public static Double haversine(Double rad) {
-    return 0.5 * (1 - Math.cos(rad));
-  }
-
-  public static String geoHash(byte[] lon, byte[] lat) {
-    Double longitude = Coder.bytesToDouble(lon);
-    Double latitude = Coder.bytesToDouble(lat);
-
-    char[] longDigits = coordDigits(longitude, -180.0, 180.0);
-    char[] latDigits = coordDigits(latitude, -90.0, 90.0);
-    char[] hashBin = new char[LEN_GEOHASH];
-    for (int c = 0; c < LEN_GEOHASH/2; c++) {
-      hashBin[2*c] = longDigits[c];
-      hashBin[(2*c)+1] = latDigits[c];
-    }
-
-    StringBuilder hashStrBuilder = new StringBuilder();
-    StringBuilder digitBuilder = new StringBuilder();
-
-    int e = 0;
-    for (int d = 0; d < LEN_GEOHASH; d++) {
-      digitBuilder.append(hashBin[d]);
-      if (e == 4) {
-        hashStrBuilder.append(base32(Integer.parseInt(digitBuilder.toString(), 2)));
-        digitBuilder = new StringBuilder();
-        e = 0;
-      } else {
-        e++;
-      }
-    }
-
-    return hashStrBuilder.toString();
-  }
-
-  private static char base32(int x) {
-    String base32str = "0123456789bcdefghjkmnpqrstuvwxyz";
-    return base32str.charAt(x);
-  }
-
-  private static int base32Val(char d) {
-    String base32str = "0123456789bcdefghjkmnpqrstuvwxyz";
-    return base32str.indexOf(d);
-  }
-
-  private static String base32bin(int v) {
-    if (v > 15) {
-      return Integer.toBinaryString(v);
-    }
-
-    if (v > 7) {
-      return "0" + Integer.toBinaryString(v);
-    }
-
-    if (v > 3) {
-      return "00" + Integer.toBinaryString(v);
-    }
-
-    if (v > 1) {
-      return "000" + Integer.toBinaryString(v);
-    }
-
-    return "0000" + Integer.toBinaryString(v);
-  }
-
-  private static char[] coordDigits(Double coordinate, Double min, Double max) {
-    char[] bin = new char[LEN_GEOHASH/2];
-    for (int c = 0; c < bin.length; c++) {
-      Double mid = (min + max) / 2;
-      if (coordinate >= mid) {
-        bin[c] = '1';
-        min = mid;
-      } else {
-        bin[c] = '0';
-        max = mid;
-      }
-    }
-
-    return bin;
-  }
-
-  private static double coord(char[] digits, Double min, Double max) {
-    Double mid = (min + max) / 2;
-    for (char d : digits) {
-      if (d == '1') {
-        min = mid;
-      } else {
-        max = mid;
-      }
-      mid = (min + max) / 2;
-    }
-
-    return mid;
   }
 
   public static ByteArrayWrapper stringToByteWrapper(String s) {
